Initial commit, customized code assembled by Toxic (Thrax989) based off Core3 by TheAnswer. Full contributions list to follow. Custom changes around Entertainer buffs, terminal buffs, XP and mob placement already implemented.

+++ MMOCoreORB/src/templates/params/creature/CreaturePosture.h
@@ -311,19 +311,19 @@ public:
 		}
 	}
 
-	int getRangedAttackMod(uint8 loc) {
+	int getRangedAttackMod(uint loc) {
 		return rangedAttackMod.get(loc);
 	}
 
-	int getRangedDefenseMod(uint8 loc) {
+	int getRangedDefenseMod(uint loc) {
 		return rangedDefenseMod.get(loc);
 	}
 
-	int getMeleeAttackMod(uint8 loc) {
+	int getMeleeAttackMod(uint loc) {
 		return meleeAttackMod.get(loc);
 	}
 
-	int getMeleeDefenseMod(uint8 loc) {
+	int getMeleeDefenseMod(uint loc) {
 		return meleeDefenseMod.get(loc);
 	}
 };
